final data collection
This a repository with all the files used in the making of the thesis " Assessing the Influence of Artificial Intelligence Adoption On Firm Performance: A Preliminary Empirical Study Using Machine Learning" by Olivier Hagoort. The general Data file used for regression analysis can be found under "DATA_variables.xlsx". The other files are used to transform, create or analysis tools. The raw data is not included due to the file being to large. The workspace attached is the where you find all the files used for the analysis of the earning calls and the creation of the dictionary.